Redo bitpack to support 128-bit integer

+++ bitpack.h
@@ -17,13 +17,13 @@
 #include <stdint.h>
 
 bool Bitpack_fitss(int64_t n, unsigned width);
-bool Bitpack_fitsu(uint64_t n, unsigned width);
+bool Bitpack_fitsu(__uint128_t n, unsigned width);
 
 int64_t Bitpack_gets(uint64_t word, unsigned width, unsigned lsb);
-uint64_t Bitpack_getu(uint64_t word, unsigned width, unsigned lsb);
+__uint128_t Bitpack_getu(__uint128_t word, unsigned width, unsigned lsb);
 
-uint64_t Bitpack_newu(uint64_t word, unsigned width, unsigned lsb,
-                      uint64_t value);
+__uint128_t Bitpack_newu(__uint128_t word, unsigned width, unsigned lsb,
+                      __uint128_t value);
 uint64_t Bitpack_news(uint64_t word, unsigned width, unsigned lsb,
                       int64_t value);
